% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_vax_rates.R
\name{cumulative_vax_rates}
\alias{cumulative_vax_rates}
\title{build cumulative vax rates tab}
\usage{
cumulative_vax_rates(width1 = 3)
}
\arguments{
\item{width1}{}
}
\value{

}
\description{
build cumulative vax rates tab
}
