% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_communities_list.R
\name{get_communities_list}
\alias{get_communities_list}
\title{Provides the list of AltaMed communities}
\usage{
get_communities_list()
}
\value{
the list of AltaMed communities
}
\description{
Provides the list of AltaMed communities
}
