% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_cumulative_vaccination_rates.R
\name{plot_community_vaccination_rates}
\alias{plot_community_vaccination_rates}
\title{Title}
\usage{
plot_community_vaccination_rates(
  community_name,
  current_community_data = data1 \%>\% filter(Community == community_name) \%>\%
    arrange(`Date (Dose 1)`)
)
}
\arguments{
\item{community_name}{}

\item{current_community_data}{}

\item{legendstats}{}
}
\value{

}
\description{
Title
}
