% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chart.R
\name{run_chart}
\alias{run_chart}
\title{Title Build run charts}
\usage{
run_chart(
  community_name,
  dataset,
  current_community_data = dataset \%>\% filter(Community == community_name) \%>\%
    arrange(`Date`)
)
}
\arguments{
\item{community_name}{}

\item{dataset}{}

\item{current_community_data}{}
}
\value{
run chart
}
\description{
Title Build run charts
}
