% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vax_by_CSA}
\alias{vax_by_CSA}
\title{Vaccination rates by CSA (combined statistical area)}
\format{
to add
}
\usage{
vax_by_CSA
}
\description{
Vaccination rates by CSA (combined statistical area)
}
\keyword{datasets}
